fix(demo-mode): use DemoUserPermission class
