+++ src/sentry/api/permissions.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any
 
-from rest_framework.permissions import BasePermission, IsAuthenticated
+from rest_framework.permissions import SAFE_METHODS, BasePermission, IsAuthenticated
 from rest_framework.request import Request
 
 from sentry.api.exceptions import (
@@ -272,6 +272,51 @@ class SentryPermission(ScopedPermission):
                     extra=extra,
                 )
                 raise MemberDisabledOverLimit(organization)
+
+
+class DemoUserPermission(SentryPermission):
+    """
+    A permission class that extends `SentryPermission` to provide read-only access for users
+    in a demo mode. This class modifies the access control logic to ensure that users identified
+    as read-only can only perform safe operations, such as GET and HEAD requests, on resources.
+    """
+
+    def determine_access(
+        self,
+        request: Request,
+        organization: RpcUserOrganizationContext | Organization | RpcOrganization,
+    ) -> None:
+
+        org_context: RpcUserOrganizationContext | None
+        if isinstance(organization, RpcUserOrganizationContext):
+            org_context = organization
+        else:
+            org_context = organization_service.get_organization_by_id(
+                id=extract_id_from(organization), user_id=request.user.id if request.user else None
+            )
+
+        if org_context is None:
+            assert False, "Failed to fetch organization in determine_access"
+
+        if demo_mode.is_demo_user(request.user):
+            if org_context.member and demo_mode.is_demo_mode_enabled():
+                org_context.member.scopes = list(demo_mode.get_readonly_scopes())
+
+        return super().determine_access(request, org_context)
+
+    def has_permission(self, request: Request, view: object) -> bool:
+        if demo_mode.is_demo_user(request.user):
+            if not demo_mode.is_demo_mode_enabled() or request.method not in SAFE_METHODS:
+                return False
+
+        return super().has_permission(request, view)
+
+    def has_object_permission(self, request: Request, view: object | None, obj: Any) -> bool:
+        if demo_mode.is_demo_user(request.user):
+            if not demo_mode.is_demo_mode_enabled() or request.method not in SAFE_METHODS:
+                return False
+
+        return super().has_object_permission(request, view, obj)
 
 
 class SentryIsAuthenticated(IsAuthenticated):
