+++ src/sentry/users/api/endpoints/user_ips.py
@@ -1,3 +1,4 @@
+from rest_framework import status
 from rest_framework.request import Request
 from rest_framework.response import Response
 
@@ -10,6 +11,7 @@ from sentry.users.api.bases.user import UserEndpoint
 from sentry.users.api.serializers.userip import UserIPSerializer
 from sentry.users.models.user import User
 from sentry.users.models.userip import UserIP
+from sentry.utils.demo_mode import is_demo_user
 
 
 @control_silo_endpoint
@@ -28,6 +30,9 @@ class UserIPsEndpoint(UserEndpoint):
 
         :auth required:
         """
+
+        if is_demo_user(user):
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
         queryset = UserIP.objects.filter(user=user)
